cleaning up return types
